+++ src/scielo/bin/xml/prodtools/data/attributes.py
@@ -6,8 +6,8 @@ from prodtools import _
 from prodtools import TABLES_PATH
 from prodtools.utils import article_utils
 from prodtools.utils.fs_utils import read_file
-from prodtools.utils.reports import validation_status
-from prodtools.utils.reports import html_reports
+from prodtools.reports import validation_status
+from prodtools.reports import html_reports
 
 
 SPS_expiration_dates = [
