+++ src/scielo/bin/xml/prodtools/reports/html_reports_collapsible.js
