+++ src/scielo/bin/xml/prodtools/reports/text_report.py
